Move Login servlet to web directory. Change Login servlet:   - Set "adminName" session attribute along with "adminId" attribute if the admin is authenticated. Change AddNewRecipe servlet:   - Remove private Static final AdminDao field.   - Remove getSession method in doGet method   - Remove getAttribute("adminId") statement in doGet method   - Remove setting session  attribute for "adminName" from adminDao.read method in doGet method. Change Dashboard servlet:   - Remove private Static final AdminDao field.   - Remove setting "adminName" session attribute by calling adminDao.read method.
