+++ src/main/java/pl/coderslab/web/Dashboard.java
@@ -2,10 +2,8 @@ package pl.coderslab.web;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import pl.coderslab.dao.AdminDao;
 import pl.coderslab.dao.PlanDao;
 import pl.coderslab.dao.RecipeDao;
-import pl.coderslab.model.LastAddedPlanDto;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,14 +12,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.util.List;
 
 @WebServlet("/app/dashboard")
 public class Dashboard extends HttpServlet {
 
     RecipeDao recipeDao = new RecipeDao();
     PlanDao planDao = new PlanDao();
-    AdminDao adminDao = new AdminDao();
     public static final Logger logger = LogManager.getLogger(IsLoggedInFilter.class);
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -31,7 +27,6 @@ public class Dashboard extends HttpServlet {
         try {
             int adminId = (int) session.getAttribute("adminId");
 
-            req.setAttribute("adminName", adminDao.read(adminId).getFirstName());
             req.setAttribute("recipeCount", recipeDao.countRecipesByAdminId(adminId));
             req.setAttribute("planCount", planDao.getPlanCountByAdminId(adminId));
             req.setAttribute("mealList", planDao.getLastAddedPlan(adminId));
